refactor: 훅 구조 개편에 따르는 변경사항 반영
